Fixed warnings, cleaned up files.

+++ Alien.h
@@ -51,7 +51,7 @@ public:
 		m_symbol [ 13 ].AsciiChar = ' ';
 		m_symbol [ 14 ].AsciiChar = ' ';
 
-		for ( size_t i = 0; i < (size_t) ( m_width * m_height ); i++ )
+		for ( int i = 0; i < (int) ( m_width * m_height ) ; i++ )
 		{
 			m_symbol [ i ].Attributes = CellColour::Fore_White;
 		}
